moved User object out of server.py. Fixes #9
